# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/characters": {
      "post": {
        "operationId": "Characters_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สร้างแล้ว",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CharacterDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST หรือ VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — สถานะปัจจุบันไม่อนุญาต",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "สร้างตัวละครของผู้ใช้ปัจจุบัน (หนึ่งคนมีได้ตัวเดียว)",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/v1/characters/current": {
      "get": {
        "operationId": "Characters_current",
        "parameters": [],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CharacterDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "ตัวละครของผู้ใช้ปัจจุบัน",
        "tags": [
          "characters"
        ]
      },
      "patch": {
        "operationId": "Characters_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCharacterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CharacterDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST หรือ VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — สถานะปัจจุบันไม่อนุญาต",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "เลือกอาชีพ — ครั้งเดียว หลังผ่านชั้น 1",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/v1/characters/current/class-trials": {
      "post": {
        "operationId": "Characters_classTrial",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassTrialDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ClassTrialDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST หรือ VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — สถานะปัจจุบันไม่อนุญาต",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "ตัวอย่างการรบของอาชีพก่อนเลือก — ตัวละครจริง เวฟเดียวกันทุกอาชีพ ไม่ได้รางวัลและไม่บันทึก",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/api/v1/region-runs": {
      "post": {
        "operationId": "RegionRuns_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegionRunDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/RegionRunDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST หรือ VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — สถานะปัจจุบันไม่อนุญาต",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "เข้าโซน — คืนคำประกาศของรอบ (EX · คู่ดวล · ล็อกเลเวล)",
        "tags": [
          "world"
        ]
      }
    },
    "/api/v1/region-runs/{id}": {
      "delete": {
        "operationId": "RegionRuns_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST หรือ VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "ออกจากโซน",
        "tags": [
          "world"
        ]
      }
    },
...
  },
  "info": {
    "title": "Code Tower — csmju-gamification-knowledge",
    "description": "ระบบย่อย Code Tower ของ CSMJU2030 · ตัวตนมาจาก Core Hub (RS256 + JWKS) · response ทุกตัวห่อด้วย envelope { success, data[, meta] } / { success: false, error }",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "coreHubBearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "ssoCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "core_hub_access_token"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "TOO_MANY_REQUESTS",
              "INTERNAL_ERROR",
              "SERVICE_UNAVAILABLE"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "description": "ใช้กับ VALIDATION_ERROR",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "ProficiencyTripleDto": {
        "type": "object",
        "properties": {
          "str": {
            "type": "number"
          },
          "int": {
            "type": "number"
          },
          "vit": {
            "type": "number"
          }
        },
        "required": [
          "str",
          "int",
          "vit"
        ]
      },
      "ProficiencyDto": {
        "type": "object",
        "properties": {
          "work": {
            "description": "งานดิบตั้งแต่เลเวลที่แล้ว",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProficiencyTripleDto"
              }
            ]
          },
          "share": {
            "description": "สัดส่วนหลังถ่วงน้ำหนัก รวม 1.0",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProficiencyTripleDto"
              }
            ]
          },
          "projectedPoints": {
            "description": "ถ้าเลเวลอัพตอนนี้จะได้ช่องละกี่แต้ม",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProficiencyTripleDto"
              }
            ]
          },
          "pointsPerLevel": {
            "type": "number"
          },
          "usingClassDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "work",
          "share",
          "projectedPoints",
          "pointsPerLevel",
          "usingClassDefault"
        ]
      },
      "StatsDto": {
        "type": "object",
        "properties": {
          "str": {
            "type": "number"
          },
          "int": {
            "type": "number"
          },
          "vit": {
            "type": "number"
          },
          "agi": {
            "type": "number"
          },
          "luk": {
            "type": "number"
          }
        },
        "required": [
          "str",
          "int",
          "vit",
          "agi",
          "luk"
        ]
      },
      "DerivedStatsDto": {
        "type": "object",
        "properties": {
          "maxHp": {
            "type": "number"
          },
          "maxMp": {
            "type": "number"
          },
          "atk": {
            "type": "number",
            "description": "พลังโจมตีกายภาพ"
          },
          "matk": {
            "type": "number",
            "description": "พลังเวท"
          },
          "def": {
            "type": "number"
          },
          "mdef": {
            "type": "number"
          },
          "speed": {
            "type": "number"
          },
          "critRate": {
            "type": "number",
            "description": "0..1"
          },
          "critDmg": {
            "type": "number",
            "description": "ตัวคูณ เช่น 1.5"
          },
          "evasion": {
            "type": "number",
            "description": "0..1"
          },
          "dropBonus": {
            "type": "number",
            "description": "0..1"
          }
        },
        "required": [
          "maxHp",
          "maxMp",
          "atk",
          "matk",
          "def",
          "mdef",
          "speed",
          "critRate",
          "critDmg",
          "evasion",
          "dropBonus"
        ]
      },
...
      "EquipmentDto": {
        "type": "object",
        "properties": {
          "weapon": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemDto"
              }
            ]
          },
          "armor": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemDto"
              }
            ]
          },
          "helmet": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemDto"
              }
            ]
          },
          "accessory": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemDto"
              }
            ]
          }
        },
        "required": [
          "weapon",
          "armor",
          "helmet",
          "accessory"
        ]
      },
      "SkillSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nameTh": {
            "type": "string"
          },
          "unlockLevel": {
            "type": "number"
          },
          "mpCost": {
            "type": "number"
          },
          "kind": {
            "type": "string",
            "enum": [
              "physical",
              "magic",
              "heal",
              "shield",
              "taunt"
            ]
          },
          "aoe": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "nameTh",
          "unlockLevel",
          "mpCost",
          "kind",
          "aoe"
        ]
      },
      "CharacterDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "classId": {
            "type": "string",
            "enum": [
              "novice",
              "warrior",
              "mage",
              "guardian"
            ]
          },
          "level": {
            "type": "number"
          },
          "exp": {
            "type": "number"
          },
          "expToNext": {
            "type": "number"
          },
          "proficiency": {
            "$ref": "#/components/schemas/ProficiencyDto"
          },
          "stats": {
            "$ref": "#/components/schemas/StatsDto"
          },
          "derived": {
            "description": "ค่าที่คำนวณแล้ว (maxHp, atk, …)",
            "allOf": [
              {
                "$ref": "#/components/schemas/DerivedStatsDto"
              }
            ]
          },
          "gold": {
            "type": "number"
          },
          "materials": {
            "type": "number"
          },
          "equipment": {
            "$ref": "#/components/schemas/EquipmentDto"
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillSummaryDto"
            }
          },
          "highestFloorCleared": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "displayName",
          "classId",
          "level",
          "exp",
          "expToNext",
          "proficiency",
          "stats",
          "derived",
          "gold",
          "materials",
          "equipment",
          "skills",
          "highestFloorCleared"
        ]
      },
      "CreateCharacterDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "example": "นักสู้หมายเลข1",
            "pattern": "^[A-Za-z0-9_฀-๿]{3,20}$"
          }
        },
        "required": [
          "displayName"
        ]
      },
      "UpdateCharacterDto": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "string",
            "enum": [
              "warrior",
              "mage",
              "guardian"
            ],
            "description": "เลือกอาชีพได้ครั้งเดียว หลังผ่านชั้น 1"
          }
        },
        "required": [
          "classId"
        ]
      },
...
      "ClassTrialResultDto": {
        "type": "object",
        "properties": {
          "victory": {
            "type": "boolean"
          },
          "wavesCleared": {
            "type": "number"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CombatEventDto"
            }
          }
        },
        "required": [
          "victory",
          "wavesCleared",
          "events"
        ]
      },
      "ClassTrialSummaryDto": {
        "type": "object",
        "properties": {
          "turns": {
            "type": "number",
            "description": "จำนวนเทิร์นที่ตัวละครได้ลงมือ"
          },
          "skillCasts": {
            "type": "number",
            "description": "ร่ายสกิลกี่ครั้ง"
          },
          "damageDealt": {
            "type": "number"
          },
          "damageTaken": {
            "type": "number"
          },
          "hpLeftPct": {
            "type": "number",
            "description": "เลือดที่เหลือตอนจบ (% ของหลอด)"
          }
        },
        "required": [
          "turns",
          "skillCasts",
          "damageDealt",
          "damageTaken",
          "hpLeftPct"
        ]
      },
      "ClassTrialDto": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "string",
            "enum": [
              "warrior",
              "mage",
              "guardian"
            ]
          },
          "level": {
            "type": "number",
            "description": "เลเวลของตัวละครที่ใช้ (เลเวลปัจจุบัน)"
          },
          "floor": {
            "type": "number",
            "description": "ชั้นของเวฟตัวอย่าง (ชั้นที่ผ่านล่าสุด อย่างน้อย 1)"
          },
          "waves": {
            "type": "number",
            "description": "จำนวนเวฟของตัวอย่าง"
          },
          "program": {
            "type": "string",
            "description": "โปรแกรมตัวอย่าง — เขียนด้วยไวยากรณ์ที่ใช้ได้ตั้งแต่ผ่านชั้น 1"
          },
          "skills": {
            "description": "สกิลที่โปรแกรมตัวอย่างเรียก (ที่ปลดแล้วที่เลเวลนี้)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxHp": {
            "type": "number"
          },
          "maxMp": {
            "type": "number"
          },
          "result": {
            "$ref": "#/components/schemas/ClassTrialResultDto"
          },
          "summary": {
            "$ref": "#/components/schemas/ClassTrialSummaryDto"
          }
        },
        "required": [
          "classId",
          "level",
          "floor",
          "waves",
          "program",
          "skills",
          "maxHp",
          "maxMp",
          "result",
          "summary"
        ]
      },
      "CreateClassTrialDto": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "string",
            "enum": [
              "warrior",
              "mage",
              "guardian"
            ],
            "description": "อาชีพที่อยากเห็นตัวอย่าง"
          }
        },
        "required": [
          "classId"
        ]
      },
...
      "EliteDto": {
        "type": "object",
        "properties": {
          "wave": {
            "type": "number"
          },
          "archetypeId": {
            "type": "string"
          },
          "nameTh": {
            "type": "string"
          }
        },
        "required": [
          "wave",
          "archetypeId",
          "nameTh"
        ]
      },
      "DuelAnnounceDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "classId": {
            "type": "string"
          },
          "level": {
            "type": "number"
          },
          "live": {
            "type": "boolean",
            "description": "true = อยู่ในโซนตอนนี้จริง · false = สแนปช็อต"
          }
        },
        "required": [
          "displayName",
          "classId",
          "level",
          "live"
        ]
      },
      "LevelSyncDto": {
        "type": "object",
        "properties": {
          "level": {
            "type": "number"
          },
          "realLevel": {
            "type": "number"
          },
          "applies": {
            "type": "boolean"
          }
        },
        "required": [
          "level",
          "realLevel",
          "applies"
        ]
      },
      "RegionRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "regionId": {
            "type": "string"
          },
          "depth": {
            "type": "number"
          },
          "floor": {
            "type": "number"
          },
          "nameTh": {
            "type": "string"
          },
          "lessonTh": {
            "type": "string"
          },
          "elite": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/EliteDto"
              }
            ]
          },
          "duel": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DuelAnnounceDto"
              }
            ]
          },
          "sync": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LevelSyncDto"
              }
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "regionId",
          "depth",
          "floor",
          "nameTh",
          "lessonTh",
          "expiresAt"
        ]
      },
      "CreateRegionRunDto": {
        "type": "object",
        "properties": {
          "regionId": {
            "type": "string",
            "example": "frostland",
            "description": "id ของภูมิภาค (จาก GET /api/v1/regions)"
          },
          "depth": {
            "type": "number",
            "minimum": 1,
            "description": "รอบที่เท่าไรของโซน (เข้าได้ถึง maxDepthAllowed)"
          }
        },
        "required": [
          "regionId",
          "depth"
        ]
      },
      "DeletedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
...
    }
  },
  "security": [
    {
      "coreHubBearer": []
    },
    {
      "ssoCookie": []
    }
  ]
}